Pass image tag into the service so containers start with the correct tag
We did not see our containers starting with the tag we were passing on
the CLI with `--tag 1.2.3`. I didn't see where that was being passed
down into the service.

I also added image and tag to the deploy DSL. By convention we run
different tags in our different environments so the ability to to tie
the tag to the environment is really useful for us.