move all arbitrary `set` calls to helper methods
Using helper methods instead of `set` calls with symbols will reduce the
number of typos of the symbols. The test and code to implement this feel
a bit echo-chambery, but it makes adding new methods to the DSL easy.

It looks like `name` is already in use in the module, so we could change
to `container_name` and that keeps parity with the `container_hostname`
method.